Updated the build to target node 8
this matches the other changes to target node 8 for this project, but does so for consumers of the package. this avoids the need to include regeneratorRuntime, because node 8 has async/await built in natively.

closes #5

BREAKING CHANGE: transpilation now targets node 8, so features that are built into node 8 do not get transpiled further. this makes this package not consumable as-is in older version of node